Add files via upload
It is a program to use of sessions for python pyramid application. It is a simple web application which shows the temporary data. When people use your web application, they frequently perform a task that requires semi-permanent data to be saved. This is called a session. Here I worked with built in sessioning support of python pyramid.