Forms using Python Pyramid framework
Store and retrieve data using the SQLAlchemy ORM atop the SQLite database. Our Pyramid-based application now needs database-backed storage of pages. This frequently means an SQL database. The Pyramid community strongly supports the SQLAlchemy project and its object-relational mapper (ORM) as a convenient, Pythonic way to interface to databases. In this application, using SQLAlchemy to a SQLite database table, we provide storage and retrieval for the pages.