SQLAlchemy and URL dispatch-based Python Pyramid application
This Pyramid application that uses technologies which related to SQL database experience. It uses SQLAlchemymechanism and URL dispatch to map URLs to code. This application with with authentication and authorization. It uses SQLite database to hold data, and SQLAlchemy to access the data in this database. This application also uses Alembic for database migrations, including initialization of the SQLite database.