Functional testing
It is an end-to-end full-stack testing using webtest. Unit tests are a common and popular approach to test-driven development. In web applications, though, the templating and entire apparatus of a web site are important parts. WebTest is a Python package that does functional testing. With WebTest we can write tests which simulate a full HTTP request against a WSGI application, then test the information in the response.