Protecting Resources With Authorization
We are assigning security statements to resources describing the permissions required to perform an operation. We will allow the other views to continue working without a password. We will introduce the Pyramid concepts of authentication, authorization, permissions, and access control lists. We assign security statements to our root resource. Then we add a permissions predicate on a view.